# api/openapi.yaml
openapi: 3.0.3
info:
  title: Counterline Terminal Access API
  version: 1.0.0
  description: |
    Terminal access only. Owner email authentication remains with Supabase.
    Serve through a same-origin /api proxy. Mutations require the exact configured
    Origin. Responses are no-store. Credentials are opaque HttpOnly, SameSite=Strict
    cookies (Secure in production), never JSON or browser storage tokens.
    Device access lasts 15 minutes. Device refresh sessions last 30 days and retain
    a 60-second single-use retry window after rotation for lost-response recovery.
    Refresh revalidates an unexpired cashier session; changed/inactive employees lock.
    Provision and refresh return only the active employee credential projection.
    Offline access is limited to seven days; manager approvals to 72 hours.
    Cashier catalog bootstrap and checkout use the terminal access plus cashier-session
    cookies. They are scoped to the terminal's store; a caller cannot choose another store.
servers:
  - url: /api
security: []
paths:
...
  /orders/push:
    post:
      operationId: pushOrder
      summary: Save one immutable order, payment, and stock movement idempotently
      security: [{ ownerBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operation_id, order, items, payment]
      responses:
        '200':
          description: Accepted order or identical replay
          content:
            application/json:
              schema:
                type: object
                required: [status, operation_id, accepted_checkpoint]
        '409': { description: Operation ID reused with another payload }
        '422': { description: Rejected validation, totals, or store reference }
        default: { $ref: '#/components/responses/Error' }
...
  /pos/orders/push:
    post:
      operationId: pushTerminalOrder
      summary: Save one cashier-terminal order idempotently
      security: [{ deviceAccess: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operation_id, order, items, payment]
      responses:
        '200': { description: Accepted order or identical replay }
        '401': { description: Device session is missing, expired, or revoked; queued sales remain local }
        '403': { description: Payload store does not match the terminal store }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    ownerBearer: { type: http, scheme: bearer, bearerFormat: Supabase JWT }
    deviceAccess: { type: apiKey, in: cookie, name: terminal_access }
    deviceRefresh: { type: apiKey, in: cookie, name: terminal_refresh }
    cashierSession: { type: apiKey, in: cookie, name: terminal_cashier }
  responses:
    Error:
      description: '400 invalid request; 401 invalid credentials/PIN; 403 origin or manager access denied; 404 entity missing in store; 429 lockout; 500 unavailable. HTTP authentication failures must not fall back to cached login.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, enum: [validation_failed, authentication_required, origin_rejected, manager_required, employee_not_found, device_not_found, pin_locked, pin_invalid, server_unavailable] }
        message: { type: string }
...